Changed the imports of the css files in components and moved them to style folder
