+++ src/App.js
@@ -2,7 +2,7 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import NavBar from "./components/NavBar";
 import CoolerListing from "./components/CoolerListing";
 import { useTheme } from "./utils/themeProvider";
-import './App.css'
+import './style/App.css'
 import CarouselComponent from "./components/CarouselComponent";
 
 function App() {
