--- src/static/css/CoolerCard.css
@@ -1,77 +1,0 @@
-/* CoolerCard.css */
-
-.card {
-  margin-top: 1.5rem;
-  border: 1px solid #ccc;
-  border-radius: 10px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  transition: transform 0.2s, box-shadow 0.2s;
-  
-}
-
-.card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 4px 8px rgba(1, 1, 1, 0.6);
-}
-
-.card-title {
-  font-size: 1.25rem;
-  margin-bottom: 0.5rem;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.card-subtitle {
-  color: #666;
-  font-size: 0.9rem;
-  margin-bottom: 0.5rem;
-}
-
-.card-text {
-  font-size: 0.9rem;
-  color: #333;
-}
-
-.list-group {
-  border-top: 1px solid #ccc;
-}
-
-.list-group-item {
-  background-color: #f8f9fa;
-  border: none;
-  padding: 0.5rem 1rem;
-  font-size: 0.85rem;
-  color: #555;
-}
-
-.list-group-item:first-child {
-  border-top: none;
-}
-
-.card-link {
-  text-decoration: none;
-  color: #007bff;
-  margin-right: 1rem;
-  font-size: 0.9rem;
-}
-
-.card-link:last-child {
-  margin-right: 0;
-}
-
-.text-muted {
-  color: #999;
-}
-
-.text-danger {
-  color: #dc3545;
-}
-
-.text-dark {
-  color: #343a40;
-}
-
-.text-warning {
-  color: #ffc107;
-}
